+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `wordPaddedDec4`.
 * Add `reversedOnto` and `commitsOntoChunks`.
 * Add `ascii(2|3|4|5|6)`.
+* Add `cstringLen` unbounded builder.
 
 ## 0.3.3.0 -- 2020-02-10